requirements.txt
fastapi
uvicorn
torch
numpy
pillow
pandas
evidently
python-multipart